Part of the code that confirms the tag system.
Was not even hard to do it. And also the whitehat site on tiktok is very bad made, and you do not offer rewards for people that make effort to make it better.
I have been on a duallity to expose this, or to report the errors and exploits. But, don´t even a reward tiktok offers, and Money shall not buy the truth that the world needs to understand!