Create Study, Report, Exploits, Comments and more.
Was not even hard to do it. And also the whitehat site on tiktok is very bad made, and you do not offer rewards for people that make effort to make it better.
I have been on a duallity to expose this, or to report the errors and exploits. But, don´t even a reward tiktok offers, and Money shall not buy the truth that the world needs to understand.

Text Available on Hindi, Spanish, English, and Portuguese.
Videos of exploiting, showing the algorithm, and the power of the big data concept with an huge I.A. Coming soon....
Expect us. And please U.S investors, government agencies, and us citizens, if you want a better for the people, all over the world, stop using it.
It needs to be remaded. And all actions to put it at leats in a better version for everyone, will be suppported by us.